requirements.txt
matplotlib==3.7.2
pandas==2.0.3
streamlit==1.25.0
gunicorn==21.2.0
altair==5.0.1
psycopg2-binary==2.9.7